requirements.txt
os
uuid
termcolor
googletrans==4.0.0-rc1
Pillow
transformers
Flask

